Backend methods for 1 word a day and previous words
Added a timer that is triggered at midnight that adds the current words of the day to the newly added `PreviousWords` list, sets the word of the day to "", then resets the timer. When the next person needs the word of the day, the `GetRandomWord` function checks to see if the word is empty or null, and if so gets and returns new word. Also added an (untested) endpoint to get the last 10 words in the `PreviousWords` list.